backend/requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# External APIs
elevenlabs==1.0.0
twilio==8.13.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3

# Redis (optional)
redis==5.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
